refactor: move forms in forms folder and use the single css file for their style
